// src/app/shared/models/turma.model.ts
export interface Turma {

    idTurma: number;
    idSerie: number;
    descricao: string;
    capacidade: number;
    dataInicio: string;
    dataFinal: string;
    idhorario: number;
    idLocal: number;
    inscritos: number;
    confirmados: number;
    ocupacaoMeta: number;
    idTurno: number;
    vagas: number;
    detalhes: string;
    descricaoSerie: string;
    descricaoTurno: string;
}